CXP-1182: fix
