+++ components/catalogs/Makefile
@@ -17,7 +17,6 @@ catalogs-tests:
 	$(MAKE) catalogs-lint-back
 	$(MAKE) catalogs-coupling-back
 	$(MAKE) catalogs-coverage-back
-	$(MAKE) catalogs-integration-back
 	$(MAKE) catalogs-acceptance-back
 	$(MAKE) catalogs-lint-front
 	$(MAKE) catalogs-unit-front
