--- components/catalogs/back/src/ServiceAPI/Event/CatalogWasCreated.php
@@ -1,22 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Akeneo\Catalogs\ServiceAPI\Event;
-
-/**
- * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-final class CatalogWasCreated
-{
-    public function __construct(
-        private string $id,
-    ) {
-    }
-
-    public function getId(): string
-    {
-        return $this->id;
-    }
-}
