+++ components/catalogs/back/src/ServiceAPI/Exception/CatalogNotFoundException.php
@@ -8,7 +8,7 @@ namespace Akeneo\Catalogs\ServiceAPI\Exception;
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class CatalogNotFoundException extends \Exception
+final class CatalogNotFoundException extends \Exception
 {
     public function __construct(
         private string $id
