+++ components/catalogs/back/src/ServiceAPI/Messenger/QueryBus.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Infrastructure\Messenger;
+namespace Akeneo\Catalogs\ServiceAPI\Messenger;
 
+use Akeneo\Catalogs\ServiceAPI\Query\Query;
 use Symfony\Component\Messenger\Exception\HandlerFailedException;
 use Symfony\Component\Messenger\HandleTrait;
 use Symfony\Component\Messenger\MessageBusInterface;
@@ -12,20 +13,28 @@ use Symfony\Component\Messenger\MessageBusInterface;
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-final class CommandBus
+final class QueryBus
 {
     use HandleTrait;
 
     public function __construct(
-        MessageBusInterface $messageBus
+        MessageBusInterface $messageBus,
     ) {
         $this->messageBus = $messageBus;
     }
 
-    public function execute(object $command): void
+    /**
+     * @template R
+     * @param \Akeneo\Catalogs\ServiceAPI\Query\Query<R> $query
+     * @return R
+     *
+     * @psalm-suppress MixedReturnStatement
+     * @psalm-suppress MixedInferredReturnType
+     */
+    public function execute(Query $query): mixed
     {
         try {
-            $this->handle($command);
+            return $this->handle($query);
         } catch (HandlerFailedException $e) {
             if (null === $e->getPrevious()) {
                 throw $e;
